refactor: Move AIProvider and Summarizer classes to separate files asdf
